some changes

+++ Calculator/Subtraction.py
@@ -1,4 +1,4 @@
 def subtraction(a, b):
-    a = int(a)
-    b = int(b)
+    a = float(a)
+    b = float(b)
     return b - a
